All good!

# Conflicts:
#	src/clearcontrol/gui/video/video2d/Stack2DDisplay.java

+++ src/clearcontrol/hardware/stages/StageDeviceBase.java
@@ -83,7 +83,7 @@ public abstract class StageDeviceBase extends VirtualDevice implements
 
 	@Override
 	public Boolean waitToBeReady(	int pIndex,
-																int pTimeOut,
+																long pTimeOut,
 																TimeUnit pTimeUnit)
 	{
 		System.out.println("waiting...");
